generic menu element
